requirements.txt
openai>=1.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
supabase>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
numpy>=1.24.0
tqdm>=4.60.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
httpx>=0.27.0
